Improved floors
I added a stone floor tile asset to the textures file, and I also
created the white tile floor texture and set it as the floor for the
entire lab. I also moved Philip's boundaries that move up and down. I
like them, but they either need to be hidden behind a wall when they go
off map, or they need to have better boundaries set.